+++ Behaviors/PassiveVehicle.cs
@@ -1,7 +1,7 @@
 #define TRACE_ADJUSTMENTS
 using UnityEngine;
 
-namespace UnitySteer.Base
+namespace UnitySteer.Behaviors
 {
 
 /// <summary>