+++ Behaviors/SteerForNeighbors.cs
@@ -1,7 +1,7 @@
 //#define DEBUG_DRAWNEIGHBORS
 using UnityEngine;
 
-namespace UnitySteer.Base
+namespace UnitySteer.Behaviors
 {
 
 /// <summary>