+++ Behaviors/SteerForSpeed.cs
@@ -1,7 +1,7 @@
 using UnityEngine;
 
 
-namespace UnitySteer.Base
+namespace UnitySteer.Behaviors
 {
 
 /// <summary>