+++ Behaviors/TickedVehicle.cs
@@ -2,7 +2,7 @@
 using UnityEngine;
 using TickedPriorityQueue;
 
-namespace UnitySteer.Base
+namespace UnitySteer.Behaviors
 {
 
 /// <summary>