+++ Editor/VehicleEditor.cs
@@ -1,6 +1,6 @@
 using UnityEditor;
 
-namespace UnitySteer.Base.Editors
+namespace UnitySteer.Behaviors.Editors
 {
 
 [CustomEditor(typeof(Vehicle))]