+++ Helpers/Vector3ToggleAttribute.cs
@@ -1,9 +1,0 @@
-﻿using UnityEngine;
-
-/// <summary>
-/// Used to indicate a Vector3 whose members are used as 0/1 toggles
-/// </summary>
-public class Vector3ToggleAttribute : PropertyAttribute 
-{
-
-}